pyproject.toml
[project]
name = "free-llm-apis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.23.0",
    "groq>=0.29.0",
    "mistralai>=1.8.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.3",
]
